distribucion personalizada hecha

+++ proyecto/src/main/java/tds/modelo/Persona.java
@@ -4,9 +4,9 @@ public interface Persona {
     String getNombre();
     void setNombre(String nombre);
     double getSaldo();
-    double getPorcentaje();
+    int getPorcentaje();
     void setSaldo(double saldo);
-    void setPorcentaje(double p);
+    void setPorcentaje(int p);
     void actualizarSaldo(double cantidad);
     void resetearSaldo();
 }
